chore: fix folder structure
