+++ lib/posthog_flutter.dart
@@ -3,4 +3,4 @@ library posthog_flutter;
 export 'src/posthog.dart';
 export 'src/posthog_observer.dart';
 export 'src/posthog_options.dart';
-export 'src/screenshot/posthog_screenshot_widget.dart';
+export 'src/replay/posthog_screenshot_widget.dart';
