+++ lib/src/replay/change_detector.dart
