+++ lib/src/replay/element_parsers/element_parser_factory.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/rendering.dart';
-import 'package:posthog_flutter/src/screenshot/element_parsers/element_parser.dart';
-import 'package:posthog_flutter/src/screenshot/element_parsers/render_image_parser.dart';
+import 'package:posthog_flutter/src/replay/element_parsers/element_parser.dart';
+import 'package:posthog_flutter/src/replay/element_parsers/image_element/position_calculator.dart';
+import 'package:posthog_flutter/src/replay/element_parsers/image_element/render_image_parser.dart';
+import 'package:posthog_flutter/src/replay/element_parsers/image_element/scaler.dart';
 
-import 'position_calculator.dart';
-import 'scaler.dart';
 
 abstract class ElementParserFactory {
   ElementParser createElementParser(Type type);
