+++ lib/src/replay/element_parsers/image_element/position_calculator.dart
