+++ lib/src/replay/element_parsers/image_element/scaler.dart
